Move texture tests into dedicated module
