+++ tests/plotting/conftest.py
@@ -72,3 +72,33 @@ def colorful_tetrahedron():
     mesh = pyvista.Tetrahedron()
     mesh.cell_data["colors"] = [[255, 255, 255], [255, 0, 0], [0, 255, 0], [0, 0, 255]]
     return mesh
+
+
+def make_two_char_img(text):
+    """Turn text into an image.
+
+    This is really only here to make a two character black and white image.
+
+    """
+    # create a basic texture by plotting a sphere and converting the image
+    # buffer to a texture
+    pl = pyvista.Plotter(window_size=(300, 300), lighting=None, off_screen=True)
+    pl.add_text(text, color='w', font_size=100, position=(0.1, 0.1), viewport=True, font='courier')
+    pl.background_color = 'k'
+    pl.camera.zoom = 'tight'
+    return pyvista.Texture(pl.screenshot()).to_image()
+
+
+@pytest.fixture()
+def cubemap(texture):
+    """Sample texture as a cubemap."""
+    return pyvista.Texture(
+        [
+            make_two_char_img('X+'),
+            make_two_char_img('X-'),
+            make_two_char_img('Y+'),
+            make_two_char_img('Y-'),
+            make_two_char_img('Z+'),
+            make_two_char_img('Z-'),
+        ]
+    )
