+++ tests/plotting/test_texture.py
@@ -164,3 +164,104 @@ def test_numpy_to_texture():
         tex = numpy_to_texture(tex_im)
     assert isinstance(tex, pv.Texture)
     assert tex.to_array().dtype == np.uint8
+
+
+@pytest.mark.parametrize('as_str', [True, False])
+@pytest.mark.parametrize('ndim', [3, 4])
+def test_save_ply_texture_array(sphere, ndim, as_str, tmpdir):
+    filename = str(tmpdir.mkdir("tmpdir").join('tmp.ply'))
+
+    texture = np.ones((sphere.n_points, ndim), np.uint8)
+    texture[:, 2] = np.arange(sphere.n_points)[::-1]
+    if as_str:
+        sphere.point_data['texture'] = texture
+        sphere.save(filename, texture='texture')
+    else:
+        sphere.save(filename, texture=texture)
+
+    mesh = pv.PolyData(filename)
+    color_array_name = 'RGB' if ndim == 3 else 'RGBA'
+    assert np.allclose(mesh[color_array_name], texture)
+
+
+@pytest.mark.parametrize('as_str', [True, False])
+def test_save_ply_texture_array_catch(sphere, as_str, tmpdir):
+    filename = str(tmpdir.mkdir("tmpdir").join('tmp.ply'))
+
+    texture = np.ones((sphere.n_points, 3), np.float32)
+    with pytest.raises(ValueError, match='Invalid datatype'):
+        if as_str:
+            sphere.point_data['texture'] = texture
+            sphere.save(filename, texture='texture')
+        else:
+            sphere.save(filename, texture=texture)
+
+    with pytest.raises(TypeError):
+        sphere.save(filename, texture=[1, 2, 3])
+
+
+def test_texture_coordinates():
+    """Test adding texture coordinates"""
+    # create a rectangle vertices
+    vertices = np.array(
+        [
+            [0, 0, 0],
+            [1, 0, 0],
+            [1, 0.5, 0],
+            [0, 0.5, 0],
+        ]
+    )
+
+    # mesh faces
+    faces = np.hstack([[3, 0, 1, 2], [3, 0, 3, 2]]).astype(np.int8)
+
+    # Create simple texture coordinates
+    t_coords = np.array([[0, 0], [1, 0], [1, 1], [0, 1]])
+    # Create the poly data
+    mesh = pv.PolyData(vertices, faces)
+    # Attempt setting the texture coordinates
+    mesh.active_t_coords = t_coords
+    # now grab the texture coordinates
+    foo = mesh.active_t_coords
+    assert np.allclose(foo, t_coords)
+    texture = pv.Texture(examples.mapfile)
+    mesh.textures['map'] = texture
+    assert mesh.textures['map'] is not None
+    mesh.clear_textures()
+    assert len(mesh.textures) == 0
+
+
+def test_multiple_texture_coordinates():
+    mesh = examples.load_airplane()
+    mesh.texture_map_to_plane(inplace=True, name="tex_a", use_bounds=False)
+    mesh.texture_map_to_plane(inplace=True, name="tex_b", use_bounds=True)
+    assert not np.allclose(mesh["tex_a"], mesh["tex_b"])
+    texture = pv.Texture(examples.mapfile)
+    mesh.textures["tex_a"] = texture.copy()
+    mesh.textures["tex_b"] = texture.copy()
+    mesh._activate_texture("tex_a")
+    assert np.allclose(mesh.active_t_coords, mesh["tex_a"])
+    mesh._activate_texture("tex_b")
+    assert np.allclose(mesh.active_t_coords, mesh["tex_b"])
+
+    # Now test copying
+    cmesh = mesh.copy()
+    assert len(cmesh.textures) == 2
+    assert "tex_a" in cmesh.textures
+    assert "tex_b" in cmesh.textures
+
+
+def test_inplace_no_overwrite_texture_coordinates():
+    mesh = pv.Box()
+    truth = mesh.texture_map_to_plane(inplace=False)
+    mesh.texture_map_to_sphere(inplace=True)
+    test = mesh.texture_map_to_plane(inplace=True)
+    assert np.allclose(truth.active_t_coords, test.active_t_coords)
+
+
+def test_activate_texture_none(grid):
+    with pytest.warns(UserWarning, match=r'not a key'):
+        assert grid._activate_texture('not a key') is None
+
+    with pytest.warns(UserWarning, match=r'No textures associated'):
+        assert grid._activate_texture(True) is None
